Extract Remote::FileBase
