+++ spec/lib/everything/blog/remote/file_base_spec.rb
@@ -1,0 +1,4 @@
+require './lib/everything/blog/remote/file_base'
+
+describe Everything::Blog::Remote::FileBase do
+end
